feat: move appearance attributes of anchor and button out of fast-foundation
