+++ packages/web-components/fast-components/temp/api-report.md
@@ -617,12 +617,12 @@ export const neutralFillHoverBehavior: import("@microsoft/fast-foundation").CSSC
 // Warning: (ae-internal-missing-underscore) The name "neutralFillInput" should be prefixed with an underscore because the declaration is marked as @internal
 //
 // @internal (undocumented)
-export const neutralFillInput: ColorRecipe<FillSwatchFamily>;
+export const neutralFillInput: import("./common").ColorRecipe<FillSwatchFamily>;
 
 // Warning: (ae-internal-missing-underscore) The name "neutralFillInputActive" should be prefixed with an underscore because the declaration is marked as @internal
 //
 // @internal (undocumented)
-export const neutralFillInputActive: ColorRecipe<string>;
+export const neutralFillInputActive: import("./common").ColorRecipe<string>;
 
 // @public
 export const neutralFillInputActiveBehavior: import("@microsoft/fast-foundation").CSSCustomPropertyBehavior;
@@ -633,7 +633,7 @@ export const neutralFillInputFocusBehavior: import("@microsoft/fast-foundation")
 // Warning: (ae-internal-missing-underscore) The name "neutralFillInputHover" should be prefixed with an underscore because the declaration is marked as @internal
 //
 // @internal (undocumented)
-export const neutralFillInputHover: ColorRecipe<string>;
+export const neutralFillInputHover: import("./common").ColorRecipe<string>;
 
 // @public
 export const neutralFillInputHoverBehavior: import("@microsoft/fast-foundation").CSSCustomPropertyBehavior;
@@ -641,7 +641,7 @@ export const neutralFillInputHoverBehavior: import("@microsoft/fast-foundation")
 // Warning: (ae-internal-missing-underscore) The name "neutralFillInputRest" should be prefixed with an underscore because the declaration is marked as @internal
 //
 // @internal (undocumented)
-export const neutralFillInputRest: ColorRecipe<string>;
+export const neutralFillInputRest: import("./common").ColorRecipe<string>;
 
 // @public
 export const neutralFillInputRestBehavior: import("@microsoft/fast-foundation").CSSCustomPropertyBehavior;
@@ -649,7 +649,7 @@ export const neutralFillInputRestBehavior: import("@microsoft/fast-foundation").
 // Warning: (ae-internal-missing-underscore) The name "neutralFillInputSelected" should be prefixed with an underscore because the declaration is marked as @internal
 //
 // @internal (undocumented)
-export const neutralFillInputSelected: ColorRecipe<string>;
+export const neutralFillInputSelected: import("./common").ColorRecipe<string>;
 
 // Warning: (ae-internal-missing-underscore) The name "neutralFillRest" should be prefixed with an underscore because the declaration is marked as @internal
 //
