+++ packages/web-components/fast-foundation/src/anchor/anchor.template.ts
@@ -7,7 +7,7 @@ import { Anchor } from "./anchor";
  * @public
  */
 export const AnchorTemplate = html<Anchor>`
-    <template class="${x => x.appearance}">
+    <template>
         <a
             class="control"
             part="control"
